working-1/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
sendgrid==6.10.0
urllib3==2.0.7
